// src/types/api/service/models/PageNewsSrchDto.ts
export interface PageNewsSrchDto {
    /**
     * 
     * @type {number}
     * @memberof PageNewsSrchDto
     */
    totalElements?: number;
    /**
     * 
     * @type {number}
     * @memberof PageNewsSrchDto
     */
    totalPages?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PageNewsSrchDto
     */
    first?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof PageNewsSrchDto
     */
    last?: boolean;
    /**
     * 
     * @type {number}
     * @memberof PageNewsSrchDto
     */
    size?: number;
    /**
     * 
     * @type {Array<NewsSrchDto>}
     * @memberof PageNewsSrchDto
     */
    content?: Array<NewsSrchDto>;
    /**
     * 
     * @type {number}
     * @memberof PageNewsSrchDto
     */
    number?: number;
    /**
     * 
     * @type {SortObject}
     * @memberof PageNewsSrchDto
     */
    sort?: SortObject;
    /**
     * 
     * @type {number}
     * @memberof PageNewsSrchDto
     */
    numberOfElements?: number;
    /**
     * 
     * @type {PageableObject}
     * @memberof PageNewsSrchDto
     */
    pageable?: PageableObject;
    /**
     * 
     * @type {boolean}
     * @memberof PageNewsSrchDto
     */
    empty?: boolean;
}